requirements.txt
aiohttp==3.7.3
aiopg==1.1.0
aiosqlite==0.16.1
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
cycler==0.10.0
discord.py==1.6.0
idna==3.1
iso8601==0.1.14
kiwisolver==1.3.1
matplotlib==3.3.4
multidict==5.1.0
numpy==1.20.1
Pillow==8.1.0
pyparsing==2.4.7
PyPika==0.44.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
yarl==1.6.3
